organized components and styles
