--- client/src/components/WindowCard.tsx
@@ -1,52 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-import styled from 'styled-components';
-
-interface IProps {
-	image: string;
-	alt: string;
-	url: string;
-}
-
-const Wrapper = styled.div`
-	width: clamp(250px, 40vw, 400px);
-	border-radius: 12px;
-	margin-bottom: 6rem;
-	box-shadow: 6px 9px 60px 8px rgba(255, 222, 173, 0.9);
-
-	&:hover {
-		cursor: pointer;
-	}
-
-	@media (min-width: 1100px) {
-		width: 20vw;
-		max-width: 450px;
-	}
-`;
-
-const Image = styled.img`
-	width: 100%;
-	max-height: 100%;
-	border-radius: 12px 12px 0 0;
-`;
-
-const TextWrapper = styled.div`
-	background-color: white;
-	border-radius: 0 0 12px 12px;
-	padding: 1rem 2rem;
-`;
-
-const WindowCard: React.FC<IProps> = ({ image, alt, url, children }) => {
-	return (
-		<Wrapper>
-			<Link to={url}>
-				<Image src={image} alt={alt} />
-				<TextWrapper>
-					<div>{children}</div>
-				</TextWrapper>
-			</Link>
-		</Wrapper>
-	);
-};
-
-export default WindowCard;
