+++ client/src/components/__test__/Button.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import Button from '../Button';
+import Button from '../UI/Button';
 
 describe('UI component tests', () => {
 	test('button displays correct input text', () => {
